requirements.txt
PyQt6>=6.7
numpy>=1.26
Pillow>=10.2
soundfile>=0.12
